feat: re-design cancel button card
